+++ assets/app.js
@@ -3,6 +3,7 @@ import "./styles/app.css";
 import { createApp } from "vue";
 import HomeComponent from "./components/HomeComponent.vue";
 import FilmComponent from "./components/FilmComponent.vue";
+import AllMoviesComponent from "./components/AllMoviesComponent.vue";
 
 const homeApp = document.getElementById("home-app");
 if (homeApp) {
@@ -12,4 +13,9 @@ if (homeApp) {
 const filmApp = document.getElementById("film-app");
 if (filmApp) {
   createApp(FilmComponent).mount("#film-app");
+}
+
+const allMoviesApp = document.getElementById("all-movies-app");
+if (allMoviesApp) {
+  createApp(AllMoviesComponent).mount("#all-movies-app");
 }